[sui-verifier] Move entry-point verifier helpers into sui-types
