+++ crates/sui-json/Cargo.toml
@@ -17,7 +17,6 @@ schemars = "0.8.12"
 sui-framework = { path = "../sui-framework" }
 sui-protocol-config = { path = "../sui-protocol-config" }
 sui-types = { path = "../sui-types" }
-sui-verifier = { path = "../sui-verifier" }
 
 move-binary-format.workspace = true
 move-bytecode-utils.workspace = true
